Tokeniser now throws an error
Namely, error 1, due to unexpected character in the input
stream. This is intentional behaviour as only a limited
amount of inputs will ever be valid. It makes sense to treat
anything not specifically mentioned as invalid - though what
it actually does right now is it considers the initial state
an error and not the initial character.

Heh.