Now tokenises test set succesfully
Test set should represent most usual possibilities. It
doesn't necessarily test for *everything* but should cover
quite a bit of scenarios.

That said, some functions I wrote to remove the repetition
of the code did remove it on a micro-level but now I'm
noticing larger patterns that are repeated several times
that could probably be simplified further but I'm not sure
of how to fit it here. It'd involve changing the structure
in such a way that regardless of mode tokens are generated
for a certain set of tokens. Moving the checks outside of
the state machine switch could possibly do this. (To avoid
things like "In MODE_TEXT we do the same thing for line
breaks as in MODE_NONE but we also add a TOKEN_TEXT" when
it'd probably be simple to just add a TOKEN_TEXT for all of
those things in there and then add an external switch or
something that specifies the correct type and then a simple
data copy from curr as the second one - and that wouldn't
even be necessary as the data is pre-set to zeroes which
would mean that I'd only have to set data[0] to curr!