Makefile fixed, parser phase started
First, I've added some arrays that look absolutely
horrendous but should make recognising keywords a tad
easier. Basically, all keywords start with letters [a-z] and
thus I created an array of all built-ins starting with a
specific letter. Because we only handle an extended set of
ASCII (some ANSI codepage, 443?) we'll follow the following
procedure:

1. Lower-case the current TOKEN_TEXT to a temporary variable
2. Check its first character char >= 'a' && char <= 'z'
3. If false, stop. This is a STOKEN_VARIABLE. Otherwise 4.
4. Stop. This is a STOKEN_KEYWORD.

In practise, STOKEN_KEYWORD will always result in
termination of the parse, because extraneous keywords are
not allowed. If we do need to allow a specific keyword,
we'll do after-testing on each STOKEN_KEYWORD and convert
it to the correct STOKEN_FUNCTION, or whatnot. Basically,
in the grammatical parse any STOKEN_KEYWORD will end the
parsing with an error. But here we can still allow it since
it's a perfectly valid *token*.